Removes unused classes for join map lane project
